requirements.txt
# Python dependencies
python-dotenv>=1.0.0
requests>=2.30.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
folium>=0.14.0
polyline>=2.0.0
ipython>=8.0.0

# Optional dependencies for advanced visualization
streamlit>=1.22.0  # For interactive web dashboard (optional)
plotly>=5.14.0     # For interactive plots (optional)
